refactor: refactor field package structure to use "masked" subpackage
