+++ webforj-foundation/src/test/java/com/webforj/component/field/masked/DwcMaskedFieldMock.java
@@ -1,4 +1,4 @@
-package com.webforj.component.field;
+package com.webforj.component.field.masked;
 
 import com.webforj.component.window.Window;
