+++ webforj-foundation/src/test/java/com/webforj/component/field/masked/MaskedNumberFieldTest.java
@@ -1,4 +1,4 @@
-package com.webforj.component.field;
+package com.webforj.component.field.masked;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
